Removed the CodeTransformation ns. Renamed the CodeTransformation class to CodeTransformationExtensions to meet conventions for classes that hold extension methods.
