+++ src/QsCompiler/Compiler/FunctorGeneration.cs
@@ -9,7 +9,7 @@ using Microsoft.Quantum.QsCompiler.DataTypes;
 using Microsoft.Quantum.QsCompiler.ReservedKeywords;
 using Microsoft.Quantum.QsCompiler.SyntaxTokens;
 using Microsoft.Quantum.QsCompiler.SyntaxTree;
-using Microsoft.Quantum.QsCompiler.Transformations.CodeTransformationsNS;
+using Microsoft.Quantum.QsCompiler.Transformations;
 
 namespace Microsoft.Quantum.QsCompiler
 {
