+++ src/QsCompiler/Transformations/CodeTransformations.cs
@@ -11,12 +11,12 @@ using Microsoft.Quantum.QsCompiler.Transformations.FunctorGeneration;
 using Microsoft.Quantum.QsCompiler.Transformations.SearchAndReplace;
 
 
-namespace Microsoft.Quantum.QsCompiler.Transformations.CodeTransformationsNS
+namespace Microsoft.Quantum.QsCompiler.Transformations
 {
     /// <summary>
     /// Static base class to accumulate the handles to individual syntax tree rewrite steps.
     /// </summary>
-    public static class CodeTransformations
+    public static class CodeTransformationExtenstions
     {
         /// <summary>
         /// Given the body of an operation, auto-generates the (content of the) adjoint specialization,
