+++ src/QsCompiler/Transformations/Conjugations.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Immutable;
 using Microsoft.Quantum.QsCompiler.SyntaxTokens;
 using Microsoft.Quantum.QsCompiler.SyntaxTree;
-using Microsoft.Quantum.QsCompiler.Transformations.CodeTransformationsNS;
 using Microsoft.Quantum.QsCompiler.Transformations.Core;
 using Microsoft.Quantum.QsCompiler.Transformations.SearchAndReplace;
